+++ src/web/src/components/Chat/Chat.jsx
@@ -262,7 +262,7 @@ const Chat = ({ state }) => {
         )}
       </Menu.Item>
     ),
-    render: () => (
+    pane: (
       <Tab.Pane
         attached={false}
         key={tab.key}
@@ -381,7 +381,7 @@ const Chat = ({ state }) => {
                 <Icon name="plus" />
               </Menu.Item>
             ),
-            render: () => null,
+            pane: null,
           },
         ]}
         renderActiveOnly={false}